requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
bson==0.5.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coverage==6.3.2
cryptography==36.0.2
Django==4.0.3
flake8==4.0.1
h11==0.13.0
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
mypy==0.941
mypy-extensions==0.4.3
mysqlclient==2.1.0
outcome==1.1.0
packaging==21.3
parameterized==0.8.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.1.1
pytest-django==4.5.2
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.1
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
types-python-slugify==5.0.4
types-requests==2.27.25
types-urllib3==1.26.14
typing_extensions==4.1.1
Unidecode==1.3.4
urllib3==1.26.9
wsproto==1.1.0
